+++ scout/adapter/mongo/variant_evaluation_terms.py
@@ -9,22 +9,13 @@ LOG = logging.getLogger(__name__)
 
 
 class VariantEvaluationHandler(object):
-    """Interact with variant evalutation information."""
+    """Interact with variant evaluation information."""
 
     def add_evaluation_term(self, evaluation_term_obj):
-        """Add evalutation term for a institute."""
+        """Add evaluation term for a institute."""
         self.evaluation_terms_collection.insert_one(evaluation_term_obj)
 
-    def update_evaluation_term(self, institute_obj, term_id, evaluation_term_obj):
-        """Update an existing term."""
-        self.evaluation_terms_collection.update()
-
-    def drop_evaluation_terms(self):
-        """Drop term collection from database"""
-        self.evaluation_terms_collection.drop()
-        LOG.info("Dropped the evaluation terms collection from database")
-
-    def evaluation_terms(self, term_category=None, analysis_type=None, institute_id=None):
+    def evaluation_terms(self, term_category=None, track=None, institute_id=None):
         """List evaluation terms used by a institute."""
         query = {}
         if term_category:
@@ -34,12 +25,12 @@ class VariantEvaluationHandler(object):
         if institute_id:
             target_institutes.append(institute_id)
 
-        target_analysis_types = ["all"]
-        if analysis_type:
-            target_analysis_types.append(analysis_type)
+        target_tracks = ["all"]
+        if track:
+            target_tracks.append(track)
 
         query = {
-            "analysis_type": {"$in": target_analysis_types},
+            "track": {"$in": target_tracks},
             "institute": {"$in": target_institutes},
             **query,
         }
@@ -47,7 +38,7 @@ class VariantEvaluationHandler(object):
         return self.evaluation_terms_collection.find(query, sort=[("rank", pymongo.ASCENDING)])
 
     def get_evaluation_term(
-        self, term_category, term_id=None, rank=None, analysis_type=None, institute_id=None
+        self, term_category, term_id=None, rank=None, track=None, institute_id=None
     ):
         """Get evaluation term data."""
         # for backwards compat
@@ -64,15 +55,15 @@ class VariantEvaluationHandler(object):
         if institute_id:
             target_institutes.append(institute_id)
 
-        target_analysis_types = ["all"]
-        if analysis_type:
-            target_analysis_types.append(analysis_type)
+        target_tracks = ["all"]
+        if track:
+            target_tracks.append(track)
 
         # build final query
         query = {
             **query,
             "term_category": term_category,
-            "analysis_type": {"$in": target_analysis_types},
+            "track": {"$in": target_tracks},
             "institute": {"$in": target_institutes},
         }
         return self.evaluation_terms_collection.find_one(query, sort=[("rank", pymongo.ASCENDING)])