+++ scout/server/blueprints/cases/controllers.py
@@ -198,10 +198,10 @@ def case(store, institute_obj, case_obj):
     omim_terms = {term["disease_nr"]: term for term in store.case_omim_diagnoses(case_obj)}
 
     # get manual rank options
-    evalutation_terms = store.evaluation_terms(
+    evaluation_terms = store.evaluation_terms(
         term_category="manual_rank", institute_id=institute_obj["internal_id"]
     )
-    manual_rank_options = build_variant_evaluation_terms(evalutation_terms)
+    manual_rank_options = build_variant_evaluation_terms(evaluation_terms)
 
     # get evaluated variants
     evaluated_variants = store.evaluated_variants(case_obj["_id"])
@@ -313,12 +313,12 @@ def case_report_content(store, institute_obj, case_obj):
 
         individual["phenotype_human"] = pheno_map.get(individual["phenotype"])
 
-    evalutation_terms = store.evaluation_terms(
+    evaluation_terms = store.evaluation_terms(
         term_category="dismissal_term",
         institute_id=institute_obj["internal_id"],
-        analysis_type="cancer" if case_obj.get("track") == "cancer" else None,
+        track="cancer" if case_obj.get("track") == "cancer" else None,
     )
-    data["dismissed_options"] = build_variant_evaluation_terms(evalutation_terms)
+    data["dismissed_options"] = build_variant_evaluation_terms(evaluation_terms)
 
     data["comments"] = store.events(institute_obj, case=case_obj, comments=True)
     data["audits"] = store.case_events_by_verb(
@@ -326,12 +326,12 @@ def case_report_content(store, institute_obj, case_obj):
     )
 
     # read manual rank options from database
-    evalutation_terms = store.evaluation_terms(
+    evaluation_terms = store.evaluation_terms(
         term_category="manual_rank",
         institute_id=institute_obj["internal_id"],
-        analysis_type="cancer" if case_obj.get("track") == "cancer" else None,
+        track="cancer" if case_obj.get("track") == "cancer" else None,
     )
-    data["manual_rank_options"] = build_variant_evaluation_terms(evalutation_terms)
+    data["manual_rank_options"] = build_variant_evaluation_terms(evaluation_terms)
 
     data["cancer_tier_options"] = CANCER_TIER_OPTIONS
     data["genetic_models"] = dict(GENETIC_MODELS)